without arguments: default to -i behavior (copy stdin)
Without arguments, we'd like to do -i only if we have some input,
else display usage info, but this check depends on both the compiler
and runtime, so we do runtime detection of the API first.

On dotnet 4.5+ the API is available and we use it. On earlier
versions the API is not available, so we do -i unconditionally.

Notably, on Windows XP only 4.0 is available, so -i is unconditional.