copy: first try Win32 clipboard API via user32.dll
Apparently the dotnet Clipboard SetText and SetDataObject can fail
repeatedly when another app monitors the clipboard (such as TightVNC
or other remote desktop apps).

This does not seem to be fixed by different methods of retry
(different retries/cooldown params with SetDataObject, or repeated
calls to the API).

However, the Win32 API seem to not fail in such cases, so now we
first try the Win32 API, and then fallback the dotnet API.

On linux, user32.dll is unavailable, so it's only the dotnet API.