paste (-o) to console: add special handling
Previously, using -o when the standard output was a (Windows) console
resulted in the output being interpreted according to the current
console (output) codepage - which is not UTF8 by default.

As a result, Unicode output looked gibberish unless the user first
set the console codepage to UTF8 (chcp 65001).

Now, -o first tries to write Unicode directly to the console using
the win32 API WriteConsoleW - which can display Unicode correctly
regardless of the console codepage, and falls back to writing UTF8.

Additionally, a new option -oo disables this special handling and
writes UTF8 bytes even if the output is a console.